# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Payments API
  version: 0.2.0
  description: Core payment processing — intents, captures, refunds, balance, and payment methods
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/payment_methods/{id}:
    get:
      summary: Get payment method by ID
      description: Returns a single payment method scoped to the authenticated merchant.
      x-platform:
        kind: resource
        entity: PaymentMethod
        affectsLedger: false
      operationId: getPaymentMethod
      tags:
        - PaymentMethods
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Payment method found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - payment_method
                properties:
                  payment_method:
                    type: object
              examples:
                found:
                  value:
                    payment_method:
                      id: cmf1pm123
                      merchant_id: cmf1merchant123
                      customer_id: cmf1customer123
                      type: card
                      status: active
                      last4: "4242"
                      brand: visa
                      created_at: 2026-03-01T10:00:00.000Z
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment method not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error:
                      code: PAYMENT_METHOD_NOT_FOUND
                      message: Payment method not found.
    patch:
      summary: Update payment method billing info
      description: >
        Updates billing details or metadata on an existing payment method. Emits `payment_method.updated`. Cannot update
        revoked methods.
      x-platform:
        kind: transition
        entity: PaymentMethod
        emitsEvents:
          - payment_method.updated
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updatePaymentMethod
      tags:
        - PaymentMethods
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                billing_name:
                  type: string
                billing_email:
                  type: string
                  format: email
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              updateBilling:
                value:
                  billing_name: Jane Updated
                  billing_email: updated@example.com
      responses:
        "200":
          description: Payment method updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - payment_method
                properties:
                  payment_method:
                    type: object
              examples:
                updated:
                  value:
                    payment_method:
                      id: cmf1pm123
                      billing_name: Jane Updated
                      billing_email: updated@example.com
                      status: active
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment method not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot update a revoked payment method.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                revoked:
                  value:
                    error:
                      code: PAYMENT_METHOD_REVOKED
                      message: Cannot update a revoked payment method.
    delete:
      summary: Detach/revoke payment method
      description: >
        Revokes a payment method, setting its status to `revoked`. Emits `payment_method.revoked`. Idempotent —
        already-revoked returns 409.
      x-platform:
        kind: transition
        entity: PaymentMethod
        stateChanges:
          - active -> revoked
          - expired -> revoked
        emitsEvents:
          - payment_method.revoked
        affectsLedger: false
      operationId: deletePaymentMethod
      tags:
        - PaymentMethods
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Payment method revoked.
          content:
            application/json:
              schema:
                type: object
                required:
                  - payment_method
                properties:
                  payment_method:
                    type: object
              examples:
                revoked:
                  value:
                    payment_method:
                      id: cmf1pm123
                      status: revoked
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment method not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Payment method is already revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/customers/{id}/payment_methods:
    get:
      summary: List customer payment methods
      description: Returns a paginated list of payment methods attached to a specific customer.
      x-platform:
        kind: resource
        entity: PaymentMethod
        affectsLedger: false
      operationId: listCustomerPaymentMethods
      tags:
        - PaymentMethods
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: type
          description: Filter by payment method type.
          schema:
            type: string
            enum:
              - card
              - bank_account
              - wallet
        - in: query
          name: status
          description: Filter by payment method status.
          schema:
            type: string
            enum:
              - active
              - expired
              - revoked
      responses:
        "200":
          description: Paginated list of customer payment methods.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
              examples:
                page:
                  value:
                    data:
                      - id: cmf1pm123
                        type: card
                        status: active
                        last4: "4242"
                        brand: visa
                    pagination:
                      page: 1
                      pageSize: 10
                      total: 1
                      hasMore: false
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/payment_intents:
    post:
      summary: Create payment intent
      description: >
        Creates a payment intent in `requires_confirmation` status and emits a `payment_intent.created` webhook delivery
        record. Requires `Idempotency-Key` header.
      x-platform:
        kind: transition
        entity: PaymentIntent
        stateChanges:
          - requires_confirmation
        emitsEvents:
          - payment_intent.created
        affectsLedger: true
        workflowRefs:
          - payment_and_refund
          - subscription_lifecycle
      operationId: createPaymentIntent
      tags:
        - PaymentIntents
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - currency
                - payment_method_token
              properties:
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                customer_id:
                  type: string
                payment_method_token:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              basic:
                value:
                  amount: 2500
                  currency: USD
                  customer_id: "{{customerId}}"
                  payment_method_token: pm_tok_visa
                  metadata:
                    order_id: ord_123
      responses:
        "201":
          description: Payment intent created in `requires_confirmation` status.
          content:
            application/json:
              schema:
                type: object
                required:
                  - payment_intent
                properties:
                  payment_intent:
                    type: object
                    properties:
                      id:
                        type: string
                      merchant_id:
                        type: string
                      customer_id:
                        type: string
                        nullable: true
                      amount:
                        type: integer
                      currency:
                        type: string
                      status:
                        type: string
                        example: requires_confirmation
                      payment_method_token:
                        type: string
                      metadata:
                        type: object
                        nullable: true
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
              examples:
                created:
                  value:
                    payment_intent:
                      id: "{{paymentIntentId}}"
                      merchant_id: cmf1merchant123
                      customer_id: "{{customerId}}"
                      amount: 2500
                      currency: USD
                      status: requires_confirmation
                      payment_method_token: pm_tok_visa
                      metadata:
                        order_id: ord_123
                      created_at: 2026-02-28T12:00:00.000Z
                      updated_at: 2026-02-28T12:00:00.000Z
        "400":
          description: Invalid payload or missing Idempotency-Key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: |
        Client-generated unique key for safe request retries. Identical key + payload replays the original response.
      schema:
        type: string
        example: idem-key-a1b2c3d4
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...